[Dy2St] Remove duplicate dy2st resnet test (#59492)
