+++ test/dygraph_to_static/CMakeLists.txt
@@ -47,7 +47,6 @@ set_tests_properties(test_bert PROPERTIES TIMEOUT 240)
 set_tests_properties(test_build_strategy PROPERTIES TIMEOUT 120)
 
 if(NOT WIN32)
-  set_tests_properties(test_resnet_v2 PROPERTIES TIMEOUT 180)
   set_tests_properties(test_tsm PROPERTIES TIMEOUT 900)
   set_tests_properties(test_resnet PROPERTIES TIMEOUT 300)
 endif()
@@ -56,7 +55,7 @@ if(APPLE)
   set_tests_properties(test_bmn PROPERTIES TIMEOUT 300)
   set_tests_properties(test_build_strategy PROPERTIES TIMEOUT 300)
   set_tests_properties(test_mobile_net PROPERTIES TIMEOUT 300)
-  set_tests_properties(test_resnet_v2 PROPERTIES TIMEOUT 300)
+  set_tests_properties(test_resnet PROPERTIES TIMEOUT 300)
 endif()
 
 if(WITH_GPU)
