+++ tools/windows/run_unittests.sh
@@ -202,7 +202,6 @@ disable_wincpu_test="^jit_kernel_test$|\
 ^test_imperative_se_resnext$|\
 ^test_bmn$|\
 ^test_mobile_net$|\
-^test_resnet_v2$|\
 ^test_build_strategy$|\
 ^test_se_resnet$|\
 ^disable_wincpu_test$"
